refactor: moved classes to separate packages
