+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/domain/subscriptions/tag/Scope.java
@@ -1,4 +1,4 @@
-package ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag;
+package ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -18,8 +18,8 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "tag_types")
-public class Type {
+@Table(name = "tag_scopes")
+public class Scope {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
