+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/dto/CommunityDTO.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.Community;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.community.Community;
 
 
 @ToString
