+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/repository/OrderRepository.java
@@ -2,7 +2,7 @@ package ru.dankoy.subscriptionsholder.subscriptions_holder.core.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Order;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
