+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/repository/ScopeRepository.java
@@ -2,7 +2,7 @@ package ru.dankoy.subscriptionsholder.subscriptions_holder.core.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Scope;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Scope;
 
 public interface ScopeRepository extends JpaRepository<Scope, Long> {
