+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/repository/SectionRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.Section;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.community.Section;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
