+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/repository/TypeRepository.java
@@ -2,7 +2,7 @@ package ru.dankoy.subscriptionsholder.subscriptions_holder.core.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Type;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Type;
 
 public interface TypeRepository extends JpaRepository<Type, Long> {
