+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/service/OrderService.java
@@ -1,7 +1,7 @@
 package ru.dankoy.subscriptionsholder.subscriptions_holder.core.service;
 
 import java.util.List;
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Order;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Order;
 
 public interface OrderService {
