+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/service/ScopeService.java
@@ -1,6 +1,6 @@
 package ru.dankoy.subscriptionsholder.subscriptions_holder.core.service;
 
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Scope;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Scope;
 
 public interface ScopeService {
