+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/service/TagService.java
@@ -1,6 +1,6 @@
 package ru.dankoy.subscriptionsholder.subscriptions_holder.core.service;
 
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Tag;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Tag;
 
 public interface TagService {
