+++ subscriptions_holder/src/main/java/ru/dankoy/subscriptionsholder/subscriptions_holder/core/service/TypeService.java
@@ -1,6 +1,6 @@
 package ru.dankoy.subscriptionsholder.subscriptions_holder.core.service;
 
-import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.tag.Type;
+import ru.dankoy.subscriptionsholder.subscriptions_holder.core.domain.subscriptions.tag.Type;
 
 public interface TypeService {
